add : umkm feat

+++ src/app/api/admin/umkm/[id]/route.ts
@@ -1,12 +1,14 @@
 import { UmkmService } from "../../../../../modules/umkm/umkm.service";
 import { ApiResponse } from "../../../../../shared/utils/response";
 import { AppError } from "../../../../../shared/errors/app-error";
+import { requireAdmin } from "../../../../../shared/auth/require-admin";
 
 export async function DELETE(
   request: Request,
   context: { params: Promise<{ id: string }> }
 ) {
   try {
+    await requireAdmin();
     const { id } = await context.params;
     await UmkmService.deleteUmkm(id);
     return ApiResponse.success(null, "UMKM berhasil dihapus");
